Refactor to allow construction of PETSc solvers with options structs
